requirements.txt
paramiko
kubernetes
pyyaml
